Move mocker in-line and add mocks for app
