+++ tests/app/authentication/test_authenticator.py
@@ -12,8 +12,11 @@ def test_check_session_with_user_id_in_session(
     session_store,
     session_data,
     expires_at,
-    mock_get_session_store,
-):  # pylint: disable=unused-argument
+    mocker,
+):
+    mocker.patch(
+        "app.authentication.authenticator.get_session_store", return_value=session_store
+    )
     with app.app_context():
         # Given
         session_store.create("eq_session_id", "user_id", session_data, expires_at)
@@ -27,8 +30,10 @@ def test_check_session_with_user_id_in_session(
         assert user.user_ik == "user_ik"
 
 
-def test_check_session_with_no_user_id_in_session(app, mock_get_session_store):
-    mock_get_session_store.return_value = None
+def test_check_session_with_no_user_id_in_session(app, mocker):
+    mocker.patch(
+        "app.authentication.authenticator.get_session_store", return_value=None
+    )
 
     with app.app_context():
         # Given / When
@@ -43,8 +48,11 @@ def test_load_user(
     session_store,
     session_data,
     expires_at,
-    mock_get_session_store,
-):  # pylint: disable=unused-argument
+    mocker,
+):
+    mocker.patch(
+        "app.authentication.authenticator.get_session_store", return_value=session_store
+    )
     with app.app_context():
         # Given
         session_store.create(
@@ -68,8 +76,11 @@ def test_request_load_user(
     session_store,
     session_data,
     expires_at,
-    mock_get_session_store,
-):  # pylint: disable=unused-argument
+    mocker,
+):
+    mocker.patch(
+        "app.authentication.authenticator.get_session_store", return_value=session_store
+    )
     with app.app_context():
         # Given
         session_store.create("eq_session_id", "user_id", session_data, expires_at)
@@ -87,8 +98,11 @@ def test_no_user_when_session_has_expired(
     app,
     session_store,
     session_data,
-    mock_get_session_store,
-):  # pylint: disable=unused-argument
+    mocker,
+):
+    mocker.patch(
+        "app.authentication.authenticator.get_session_store", return_value=session_store
+    )
     with app.app_context():
         # Given
         session_store.create(
@@ -112,8 +126,11 @@ def test_valid_user_does_not_extend_session_expiry_when_expiry_less_than_60_seco
     session_store,
     session_data,
     expires_at,
-    mock_get_session_store,
-):  # pylint: disable=unused-argument
+    mocker,
+):
+    mocker.patch(
+        "app.authentication.authenticator.get_session_store", return_value=session_store
+    )
     with app.app_context():
         # Given
         session_store.create("eq_session_id", "user_id", session_data, expires_at)
@@ -135,8 +152,11 @@ def test_valid_user_extends_session_expiry_when_expiry_greater_than_60_seconds_d
     session_store,
     session_data,
     expires_at,
-    mock_get_session_store,
-):  # pylint: disable=unused-argument
+    mocker,
+):
+    mocker.patch(
+        "app.authentication.authenticator.get_session_store", return_value=session_store
+    )
     with app.app_context():
         # Given
         session_store.create("eq_session_id", "user_id", session_data, expires_at)
